Update sensors migration to use float type for ph, suhu, tds, and kualitas fields

+++ database/migrations/2025_02_26_095814_create_sensors_table.php
@@ -13,10 +13,10 @@ return new class extends Migration
     {
         Schema::create('sensors', function (Blueprint $table) {
             $table->id();
-            $table->double('ph');
-            $table->double('suhu');
-            $table->double('tds');
-            $table->double('kualitas');
+            $table->float('ph');
+            $table->float('suhu');
+            $table->float('tds');
+            $table->float('kualitas');
             $table->timestamps();
         });
     }
